backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-jose[cryptography]==3.3.0
boto3>=1.34.129
passlib[bcrypt]==1.7.4
requests-oauthlib>=2.0.0
python-multipart==0.0.9
cryptography>=42.0.8
python-dotenv==1.0.1
sqlalchemy==2.0.28
pymongo==4.5.0
alembic==1.13.1
pyjwt>=2.10.1
requests==2.31.0
pytest==8.0.0
pytest-asyncio>=0.23.7
aiohttp>=3.10.5
tzdata>=2024.2
httpx==0.27.0
motor==3.3.1
black==24.2.0
isort==5.13.2
mypy==1.8.0
flake8>=7.0.0
pandas>=2.2.0
numpy>=1.26.0
jq>=1.6.0
typer>=0.9.0

# Extra useful packages (avoid duplicates)
pydantic==2.11.9
pydantic-settings>=2.2.1
